# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Brazino Accs - Roblox account manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Account {
  id: number;
  username: string;
  /** @nullable */
  password?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  authenticator?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface AccountInput {
  /** @minLength 1 */
  username: string;
  password?: string;
  email?: string;
  authenticator?: string;
  avatarUrl?: string;
  notes?: string;
}

export interface AccountPatch {
  /** @minLength 1 */
  username?: string;
  password?: string;
  email?: string;
  authenticator?: string;
  avatarUrl?: string;
  notes?: string;
}

export interface AccountStats {
  total: number;
  withEmail: number;
  withAuthenticator: number;
  withPassword: number;
  recent: number;
}