# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Brazino Accs - Roblox account manager API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: accounts
    description: Roblox account operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List all saved accounts
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Save a new account (called by extension)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountInput"
      responses:
        "201":
          description: Account saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          description: Invalid input

  /accounts/stats:
    get:
      operationId: getAccountStats
      tags: [accounts]
      summary: Get account statistics
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountStats"

  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get a single account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Not found
    patch:
      operationId: updateAccount
      tags: [accounts]
      summary: Update an account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Not found
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Delete an account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Account:
      type: object
      required: [id, username, createdAt, updatedAt]
      properties:
        id:
          type: integer
        username:
          type: string
        password:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        authenticator:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    AccountInput:
      type: object
      required: [username]
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
        email:
          type: string
        authenticator:
          type: string
        avatarUrl:
          type: string
        notes:
          type: string

    AccountPatch:
      type: object
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
        email:
          type: string
        authenticator:
          type: string
        avatarUrl:
          type: string
        notes:
          type: string

    AccountStats:
      type: object
      required: [total, withEmail, withAuthenticator, withPassword, recent]
      properties:
        total:
          type: integer
        withEmail:
          type: integer
        withAuthenticator:
          type: integer
        withPassword:
          type: integer
        recent:
          type: integer