# lib/api-zod/src/generated/types/account.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Brazino Accs - Roblox account manager API
 * OpenAPI spec version: 0.1.0
 */

export interface Account {
  id: number;
  username: string;
  /** @nullable */
  password?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  authenticator?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
  updatedAt: Date;
}