# lib/api-zod/src/generated/types/accountInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Brazino Accs - Roblox account manager API
 * OpenAPI spec version: 0.1.0
 */

export interface AccountInput {
  /** @minLength 1 */
  username: string;
  password?: string;
  email?: string;
  authenticator?: string;
  avatarUrl?: string;
  notes?: string;
}